% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodeplot.R
\name{bodeplot}
\alias{bodeplot}
\title{Bode plot}
\usage{
bodeplot(fr, fr.unit = NULL, n = NULL, d = NULL, A = NULL, B = NULL,
  C = NULL, D = NULL, highlight = NULL, label = NULL, as.period = F)
}
\arguments{
\item{fr}{Frequency range in powers of 10}

\item{fr.unit}{Unit of the frequency}

\item{n}{Numerator coefficients of a transfer function}

\item{d}{Denominator coefficients of a transfer function}

\item{A}{Matrix A of a state-space representation}

\item{B}{Matrix B of a state-space representation}

\item{C}{Matrix C of a state-space representation}

\item{D}{Matrix D of a state-space representation}

\item{highlight}{Optional highlights in the plot}

\item{label}{Optional labels on the highlights}

\item{as.period}{Flag to plot as period}
}
\value{
Bode magnitude and phase plot
}
\description{
Bode plot
}
\examples{
bodeplot(fr = c(0, 1), n = c(1, 0.1, 7.5), d = c(1, 0.12, 9, 0, 0))
bodeplot(fr = c(0, 1),
          A = matrix(c(-1,1,0,-1), nrow = 2),
          B = matrix(c(-1,1), nrow = 2),
          C = matrix(c(0,1), ncol=2),
          D = 0,
          highlight = 2, label = "Highlight")

}

