% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyquistplot.R
\name{nyquistplot}
\alias{nyquistplot}
\title{Nyquist plot}
\usage{
nyquistplot(fr, n = NULL, d = NULL)
}
\arguments{
\item{fr}{Frequency range in powers of 10}

\item{n}{Numerator coefficients of a transfer function}

\item{d}{Denominator coefficients of a transfer function}
}
\value{
Nyquist plot
}
\description{
Nyquist plot
}
\examples{
nyquistplot(c(-2, 3), n = c(2, 5, 1), d = c(1, 2, 3))
}

