% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precis.R
\name{precis}
\alias{precis}
\title{Produces a precis of variables}
\usage{
precis(
  df,
  ...,
  group_var = NA,
  types = c("n", "nobs", "nmiss", "mean", "sd", "min", "p25", "median", "p75", "max"),
  long = F,
  na.rm = T
)
}
\arguments{
\item{df}{a dataframe containing the variables to summarize}

\item{...}{one or more variables to summarize}

\item{group_var}{if given, statistics will be computed for each level of groupby variable}

\item{types}{types of summary statistics to computed (as vector of strings). Default is all of the following: "n", "nobs", "nmiss", "sum","mean","sd", "min", "p1", "p5", "p25", "median", "p75", "p95", "p99", "max"}

\item{long}{if TRUE (default), aranges statistics in columns and variable in rows Otherwise, shows reverse}

\item{na.rm}{whether or not na.rm==T is passsed to summarizing functions.}
}
\value{
Returns a dataframe containing summary statistics.
}
\description{
Produces a precis of variables
}
\examples{
\dontrun{}
precis(df, mpg, cyl)

}
